requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
streamlit
python-dotenv
google-generativeai
requests
langchain
langchain-google-genai
langchain-community